making cmake prettier and stisfying clang++

+++ src/Responce/Responce.cpp
@@ -2,7 +2,7 @@
 
 Responce::Responce() : Responce("") {}
 Responce::Responce(std::string data) : Message(data) {}
-Responce::Responce(std::array<char, BUFFER_SIZE> data, int length) :
+Responce::Responce(std::array<char, BUFFER_SIZE> data, size_t length) :
     Responce(std::string(data.data(), length)){}
 
 Responce::~Responce() {}
